+++ functions.php
@@ -24,7 +24,7 @@ require_once get_template_directory() . '/lib/init.php'; // Initialize Genesis F
  * Add or remove files to the array as needed.
  * Please note that missing files will produce a fatal error.
  *
- * @since 3.2.2
+ * @since 3.3.0
  */
 $genesis_advanced_includes = [
 
@@ -43,7 +43,7 @@ $genesis_advanced_includes = [
 	'lib/extras.php', // Custom functions.
 
 	/* Gutenberg Blocks */
-	'lib/blocks/setup.php', // Gutenberg blocks theme support.
+	'lib/blocks/blocks-setup.php', // Gutenberg blocks theme support.
 
 	/* Structure */
 	'lib/structure/content.php',